@@ -4,67 +4,63 @@
   "description": "homeautomation",
   "author": "florian hoffmann",
   "license": "MIT",
+  "sideEffects": false,
   "dependencies": {
     "axios": "^0.18.0",
     "babel-polyfill": "^6.26.0",
-    "bluebird": "^3.5.1",
-    "body-parser": "^1.18.2",
+    "bluebird": "^3.5.2",
+    "body-parser": "^1.18.3",
     "chart.js": "^2.7.2",
-    "docxtemplater": "3.6.2",
+    "compression": "^1.7.3",
+    "docxtemplater": "3.9.0",
     "g": "^2.0.1",
     "int-to-rgb": "^1.0.0",
     "jszip": "2.6.1",
     "miio": "0.15.6",
-    "mongodb": "^3.0.8",
-    "npm": "^6.0.1",
-    "office-ui-fabric-react": "^5.100.0",
-    "react": "16.3.2",
-    "react-chartjs-2": "^2.7.2",
-    "react-dom": "16.3.2",
-    "react-router": "^4.2.0",
-    "react-router-dom": "^4.2.2",
+    "mongodb": "^3.1.4",
+    "office-ui-fabric-react": "^6.67.2",
+    "react": "16.5.0",
+    "react-chartjs-2": "^2.7.4",
+    "react-dom": "16.5.0",
+    "react-router": "^4.3.1",
+    "react-router-dom": "^4.3.1",
     "serve-favicon": "^2.5.0",
-    "typescript": "^2.8.3",
-    "webpack": "^4.8.3"
+    "typescript": "^3.0.3",
+    "uuid": "^3.3.2"
   },
   "devDependencies": {
     "@types/axios": "^0.14.0",
-    "@types/bluebird": "^3.5.20",
-    "@types/chart.js": "^2.7.18",
-    "@types/express": "4.11.1",
-    "@types/jszip": "3.1.3",
+    "@types/bluebird": "^3.5.24",
+    "@types/chart.js": "^2.7.33",
+    "@types/express": "4.16.0",
+    "@types/jszip": "3.1.4",
     "@types/microsoft-ajax": "0.0.33",
-    "@types/mongodb": "^3.0.18",
-    "@types/node": "^10.0.8",
-    "@types/react": "^16.3.14",
+    "@types/mongodb": "^3.1.7",
+    "@types/node": "^10.9.4",
+    "@types/react": "^16.4.14",
     "@types/react-chartjs-2": "^2.5.7",
-    "@types/react-dom": "16.0.5",
-    "@types/react-router": "^4.0.25",
-    "@types/react-router-dom": "^4.2.6",
+    "@types/react-dom": "16.0.7",
+    "@types/react-router": "^4.0.30",
+    "@types/react-router-dom": "^4.3.0",
     "@types/requirejs": "2.1.31",
-    "@types/uuid": "^3.4.3",
-    "@types/webpack": "4.1.6",
-    "awesome-typescript-loader": "^5.0.0",
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^7.1.4",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
-    "compression-webpack-plugin": "^1.1.11",
-    "copy-webpack-plugin": "4.5.1",
-    "css-loader": "^0.28.11",
-    "expose-loader": "^0.7.5",
+    "@types/uuid": "^3.4.4",
+    "awesome-typescript-loader": "^5.2.1",
+    "clean-webpack-plugin": "^0.1.19",
+    "compression-webpack-plugin": "^2.0.0",
+    "copy-webpack-plugin": "4.5.2",
+    "css-loader": "^1.0.0",
     "express": "^4.16.3",
-    "file-loader": "1.1.11",
-    "source-map-loader": "^0.2.3",
-    "style-loader": "^0.21.0",
-    "ts-loader": "4.3.0",
-    "tslint": "^5.10.0",
+    "fork-ts-checker-webpack-plugin": "^0.4.9",
+    "hard-source-webpack-plugin": "^0.12.0",
+    "source-map-loader": "^0.2.4",
+    "style-loader": "^0.23.0",
+    "thread-loader": "1.2.0",
+    "ts-loader": "5.1.0",
+    "tslint": "^5.11.0",
     "tslint-loader": "^3.6.0",
-    "url-loader": "^1.0.1",
-    "uuid": "^3.2.1",
-    "webpack-bundle-analyzer": "^2.11.2",
-    "webpack-cli": "^2.1.3",
+    "webpack": "4.18.0",
+    "webpack-bundle-analyzer": "^3.0.1",
+    "webpack-cli": "^3.1.0",
     "webpack-notifier": "^1.6.0"
   },
   "scripts": {
