@@ -1,3 +1,5 @@
 export enum PageType {
-    Display = 0, Edit = 1, Add = 2
+  Display = 0,
+  Edit = 1,
+  Add = 2
 }
